# Fix timer usage in benchmarks